requirements.txt
Flask==2.0.1
Pillow==8.3.0
pyTelegramBotAPI==3.8.1
python-dateutil==2.8.1
qrcode==7.0
gunicorn==20.1.0
